# docs/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "infra-orchestrator API",
    "version": "0.1.0",
    "description": "Authenticated infrastructure control plane. All remote mutations pass through RBAC, policy and the operation queue. Responses never expose stored credentials."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "session": []
    },
    {
      "bearer": []
    }
  ],
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "post__api_v1_auth_login",
        "summary": "POST /api/v1/auth/login",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Successful query or mutation"
          },
          "400": {
            "description": "Invalid structured input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "RBAC, policy, MFA, environment or CSRF denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "description": "Required for cookie authentication; value from io_csrf cookie.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/config": {
      "get": {
        "operationId": "get__api_v1_auth_config",
        "summary": "GET /api/v1/auth/config",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Successful query or mutation"
          },
          "400": {
            "description": "Invalid structured input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "RBAC, policy, MFA, environment or CSRF denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/oidc/start": {
      "get": {
        "operationId": "get__api_v1_auth_oidc_start",
        "summary": "GET /api/v1/auth/oidc/start",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Successful query or mutation"
          },
          "400": {
            "description": "Invalid structured input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "RBAC, policy, MFA, environment or CSRF denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/oidc/callback": {
      "get": {
        "operationId": "get__api_v1_auth_oidc_callback",
        "summary": "GET /api/v1/auth/oidc/callback",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Successful query or mutation"
          },
          "400": {
            "description": "Invalid structured input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "RBAC, policy, MFA, environment or CSRF denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "get__api_v1_auth_me",
        "summary": "GET /api/v1/auth/me",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Successful query or mutation"
          },
          "400": {
            "description": "Invalid structured input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "RBAC, policy, MFA, environment or CSRF denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "post__api_v1_auth_logout",
        "summary": "POST /api/v1/auth/logout",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Successful query or mutation"
          },
          "400": {
            "description": "Invalid structured input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "RBAC, policy, MFA, environment or CSRF denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "description": "Required for cookie authentication; value from io_csrf cookie.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "description": "Endpoint-specific structured fields; see module documentation and source request type."
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/renew": {
      "post": {
        "operationId": "post__api_v1_auth_renew",
        "summary": "POST /api/v1/auth/renew",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Successful query or mutation"
          },
          "400": {
            "description": "Invalid structured input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "RBAC, policy, MFA, environment or CSRF denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "description": "Required for cookie authentication; value from io_csrf cookie.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "description": "Endpoint-specific structured fields; see module documentation and source request type."
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password": {
      "post": {
        "operationId": "post__api_v1_auth_password",
        "summary": "POST /api/v1/auth/password",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Successful query or mutation"
          },
          "400": {
            "description": "Invalid structured input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "RBAC, policy, MFA, environment or CSRF denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "description": "Required for cookie authentication; value from io_csrf cookie.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "description": "Endpoint-specific structured fields; see module documentation and source request type."
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/enroll": {
      "post": {
        "operationId": "post__api_v1_auth_mfa_enroll",
        "summary": "POST /api/v1/auth/mfa/enroll",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Successful query or mutation"
          },
          "400": {
            "description": "Invalid structured input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "RBAC, policy, MFA, environment or CSRF denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "description": "Required for cookie authentication; value from io_csrf cookie.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "description": "Endpoint-specific structured fields; see module documentation and source request type."
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/verify": {
      "post": {
        "operationId": "post__api_v1_auth_mfa_verify",
        "summary": "POST /api/v1/auth/mfa/verify",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Successful query or mutation"
          },
          "400": {
            "description": "Invalid structured input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "RBAC, policy, MFA, environment or CSRF denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "description": "Required for cookie authentication; value from io_csrf cookie.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "description": "Endpoint-specific structured fields; see module documentation and source request type."
              }
            }
          }
        }
      }
    },
    "/api/v1/users": {
      "get": {
        "operationId": "get__api_v1_users",
        "summary": "GET /api/v1/users",
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "Successful query or mutation"
          },
          "400": {
            "description": "Invalid structured input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "RBAC, policy, MFA, environment or CSRF denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      },
      "post": {
        "operationId": "post__api_v1_users",
        "summary": "POST /api/v1/users",
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "Successful query or mutation"
          },
          "400": {
            "description": "Invalid structured input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "RBAC, policy, MFA, environment or CSRF denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "description": "Required for cookie authentication; value from io_csrf cookie.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "description": "Endpoint-specific structured fields; see module documentation and source request type."
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{id}": {
      "put": {
        "operationId": "put__api_v1_users__id_",
        "summary": "PUT /api/v1/users/{id}",
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "Successful query or mutation"
          },
          "400": {
            "description": "Invalid structured input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "RBAC, policy, MFA, environment or CSRF denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "description": "Required for cookie authentication; value from io_csrf cookie.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "description": "Endpoint-specific structured fields; see module documentation and source request type."
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete__api_v1_users__id_",
        "summary": "DELETE /api/v1/users/{id}",
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "Successful query or mutation"
          },
          "400": {
            "description": "Invalid structured input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "RBAC, policy, MFA, environment or CSRF denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "description": "Required for cookie authentication; value from io_csrf cookie.",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/users/{id}/reset-password": {
      "post": {
        "operationId": "post__api_v1_users__id__reset_password",
        "summary": "POST /api/v1/users/{id}/reset-password",
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "Successful query or mutation"
          },
          "400": {
            "description": "Invalid structured input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "RBAC, policy, MFA, environment or CSRF denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "description": "Required for cookie authentication; value from io_csrf cookie.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "description": "Endpoint-specific structured fields; see module documentation and source request type."
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions": {
      "get": {
        "operationId": "get__api_v1_sessions",
        "summary": "GET /api/v1/sessions",
        "tags": [
          "sessions"
        ],
        "responses": {
          "200": {
            "description": "Successful query or mutation"
          },
          "400": {
            "description": "Invalid structured input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "RBAC, policy, MFA, environment or CSRF denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "io_session"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "Login": {
        "type": "object",
        "required": [
          "login",
          "password"
        ],
        "properties": {
          "login": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          },
          "code": {
            "type": "string",
            "pattern": "^[0-9]{6}$"
          },
          "method": {
            "type": "string",
            "enum": [
              "local",
              "ldap"
            ]
          }
        }
      },
...
    }
  }
}